requirements.txt
beautifulsoup4==4.11.1
cloudscraper==1.2.71
Flask==3.0.2
flask_cors==4.0.0
pandas==2.2.1
plotly==5.16.1
pymongo==4.6.0
python_dateutil==2.8.2
Requests==2.31.0
